# planner_cspace: implement motion primitive algorithm for speed-up