--- a/planner_cspace/include/planner_cspace/planner_3d/grid_astar_model.h
+++ b/planner_cspace/include/planner_cspace/planner_3d/grid_astar_model.h
@@ -35,12 +35,12 @@
 
 #include <costmap_cspace_msgs/MapMetaData3D.h>
 
-#include <planner_cspace/cyclic_vec.h>
-#include <planner_cspace/planner_3d/motion_cache.h>
-#include <planner_cspace/planner_3d/rotation_cache.h>
-#include <planner_cspace/planner_3d/path_interpolator.h>
-#include <planner_cspace/grid_astar_model.h>
 #include <planner_cspace/blockmem_gridmap.h>
+#include <planner_cspace/cyclic_vec.h>
+#include <planner_cspace/grid_astar_model.h>
+#include <planner_cspace/planner_3d/motion_cache.h>
+#include <planner_cspace/planner_3d/path_interpolator.h>
+#include <planner_cspace/planner_3d/rotation_cache.h>
 
 namespace planner_cspace
 {
@@ -81,7 +81,7 @@ protected:
   costmap_cspace_msgs::MapMetaData3D map_info_;
   Vecf euclid_cost_coef_;
   Vecf resolution_;
-  std::vector<Vec> search_list_;
+  std::vector<std::vector<Vec>> motion_primitives_;
   std::vector<Vec> search_list_rough_;
   int local_range_;
   BlockMemGridmapBase<float, 3, 2>& cost_estim_cache_;