--- a/planner_cspace/include/planner_cspace/planner_3d/motion_primitive_builder.h
+++ b/planner_cspace/include/planner_cspace/planner_3d/motion_primitive_builder.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018-2019, the neonavigation authors
+ * Copyright (c) 2020, the neonavigation authors
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without