Instead of whole search_list, only followable moving primitives are remained in A* algorithm.
This PR can reduce calculation time significantly.

The following performance comparisons are performed using our simulator.
Times are obtained from messages shown by L1547 of planner_3d.cpp.

Average time while 3-dof planning: 
master: 0.3514 sec.
This PR: 0.0520 sec.
Worst time: 
master: 0.9421 sec.
This PR: 0.1582 sec.