@at-wat 2020-01-23
## [[#1246](https://travis-ci.org/at-wat/neonavigation/builds/640850301)] FAILED on kinetic

```catkin_make``` failed