@at-wat 2020-01-23
## [[#1247](https://travis-ci.org/at-wat/neonavigation/builds/640857712)] FAILED on melodic

```catkin_make tests``` failed